+++ org.mixedrealitytoolkit.spatialmanipulation/ObjectManipulator/ObjectManipulator.cs
@@ -762,7 +762,11 @@ namespace MixedReality.Toolkit.SpatialManipulation
                 {
                     if (interactorsSelecting.Count == 0)
                     {
+#if UNITY_6000_0_OR_NEWER
+                        ReleaseRigidBody(rigidBody.linearVelocity, rigidBody.angularVelocity);
+#else
                         ReleaseRigidBody(rigidBody.velocity, rigidBody.angularVelocity);
+#endif
                     }
                     else
                     {
@@ -910,7 +914,11 @@ namespace MixedReality.Toolkit.SpatialManipulation
             // when player is moving, we need to anticipate where the targetTransform is going to be one time step from now
             distance -= referenceFrameVelocity * Time.fixedDeltaTime;
 
+#if UNITY_6000_0_OR_NEWER
+            var velocity = rigidBody.linearVelocity;
+#else
             var velocity = rigidBody.velocity;
+#endif
 
             var acceleration = omega * omega * -distance;  // acceleration caused by spring force
 
@@ -939,7 +947,11 @@ namespace MixedReality.Toolkit.SpatialManipulation
 
             velocity += referenceFrameVelocity;  // change back to global frame of reference
 
+#if UNITY_6000_0_OR_NEWER
+            rigidBody.linearVelocity = velocity;
+#else
             rigidBody.velocity = velocity;
+#endif
 
             if (applyTorque)
             {
@@ -1037,7 +1049,11 @@ namespace MixedReality.Toolkit.SpatialManipulation
                 {
                     if (releaseBehavior.IsMaskSet(ReleaseBehaviorType.KeepVelocity))
                     {
+#if UNITY_6000_0_OR_NEWER
+                        rigidBody.linearVelocity = velocity;
+#else
                         rigidBody.velocity = velocity;
+#endif
                     }
 
                     if (releaseBehavior.IsMaskSet(ReleaseBehaviorType.KeepAngularVelocity))