+++ org.mixedrealitytoolkit.spatialmanipulation/CHANGELOG.md
@@ -2,8 +2,6 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## Unreleased
-
 ## [4.0.0-development.pre.1] - 2024-07-09
 
 ### Added
@@ -18,7 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 * Updated package com.unity.xr.interaction.toolkit to 3.0.4
 
-## [3.3.0] - 2024-04-30
+## Unreleased
 
 ### Fixed