+++ org.mixedrealitytoolkit.spatialmanipulation/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [4.0.0-development.pre.2] - 2024-10-28
+
+### Changed
+
+* Updated ObjectManipulator and ObjectManipulatorTests to be compatible with renamed rigidbody properties in Unity 6.
+
 ## [4.0.0-development.pre.1] - 2024-07-09
 
 ### Added